Create IN Darkness Game
The story of the game is that you were sent into space to search for life on other planets, and you were surprised by the presence of life of advanced and different monsters on the surface of a dark planet, but unfortunately for you, the base’s engine was damaged, and that made you imprisoned on that planet. Fortunately for you, that planet has oxygen for breathing and is full of minerals that... Flatten through which to develop your base and survive.